Post Editor: pass correct value of hasContent to components
The correct value needs to be a combination of Flux content (`state.hasContent` coming
from Flux store) and Redux content (`props.hasContent` coming from a Redux selector).
This patch fixes that.

The visible result is that the `QuickSaveButtons` and Publish Button are enabled when
there is only a title and no content: `hasContent` is true in that case.