Move calls to `translate` to within the `render` method
As part of a similar change to the Medium importer, it was suggested
that the calls to `translate` should be moved within the render method
so that the strings were updated on each render of the component,
rather than translated once as the object was instantiated.

https://github.com/Automattic/wp-calypso/pull/24350#discussion_r188314191

This brings the creation of the `importer_data` object within `render`.