Fix for #4099 (geotagging not working, location doesn't display on front-end) (#24905)
Previously, posts and pages that were geo-tagged in Calypso did not save geo_public and geo_address meta values.  This resulted in geo-tagging data not rendering on WP.com.  This commit addresses these 2 missing meta values.

For geo_public, we always set it to true when updating the geo-tagging data.  This matches up with the behavior of the equivalent wp-admin UI on WP.com.  However, because geo-tagging previously saved with Calypso was de facto private due to the lack of a geo_public value, we take care to display an additional notice in the UI when a user edits the location on one of these "private" geo-tagging posts (i.e. "Note: The location will be displayed publicly.").

For geo_address, we already had access to the "formatted_address" field of the Google Maps API response in the context of selecting a result from the address search.  However, when using the browser's geo-location API, we previously only had the latitude and longitude.  With this commit, we now do a "reverse geo-code" to retrieve a human-readable formatted_address from Google following the use of the browser geo-location API.

When setting geo_address, we also render that formatted_address in the search input.  This allows us to manage this additional meta value without further cluttering the sidebar UI.  This does make the Search component responsible for both searching and displaying, which is handled by ignoring "searches" that are actually just a side effect of setting the value to the geo_address string.

With those two meta values saved, geo-tagging data from Calypso should now behave equivalently to geo-tagging data saved in wp-admin's UI.