Settings: redirect to subsection of the traffic setion when arriving from elsewhere in Calypso.
Atm sections are marked with hashtags using html5 id keys. When redirected from elsewhere within Calypso
the hashtag will not be read and thus the top of the page is set in view no matter the section passed via url.
This commit expands on the definition of the redirect-to-traffic-section function to redirect to relevant page
subsection after loading the page.
Also, href link from the Google Analytics feature card is adapted to point to the GA contentin the /traffic tab.