Post Editor: show correct status label for new draft (#25036)
* Post Editor: show correct status label for new draft

Fixes the `EditorStatusLabel` component to display a "New Draft" label
instead of "Loading Post..." for new drafts.

If the `post` is `null`, it can mean either that the post is not loaded yet,
or that it's a new draft that hasn't been saved yet.

Use the `isEditorNewPost` selector to distinguish between these cases. That's
more reliable than `post.modified`, too.

* Post Editor: pass correct value of hasContent to components

The correct value needs to be a combination of Flux content (`state.hasContent` coming
from Flux store) and Redux content (`props.hasContent` coming from a Redux selector).
This patch fixes that.

The visible result is that the `QuickSaveButtons` and Publish Button are enabled when
there is only a title and no content: `hasContent` is true in that case.