Add option for Blogger Importer (#25312)
* Add config for enabling/disabling the blogger.com importer feature

- Blogger.com importer is enabled in development environments only

* Added required constant for the blogger importer

* Add new importer-blogger component for blogger.com imports

This component very closely mirrors the existing importer-medium
component, with only a few alterations

* Add new importer-blogger component to the site importer

This commit introduces the blogger import functionality, but contains
some code that can be improved. Notably the boolean logic around whether
individual importers are rendered.

* Add propTypes to importer-blogger

- Add missing propType for `site` prop
- Add propTypes for `importerStatus` which are passed through to the
FileImporter component

* Make `importerData` a class property in blogger importer

- Make `importerData` a class proprerty
- Avoid the mutation of this object in the render method - define all of
the properties for this object in the same place to reduce context misdirection

Changes made based on code review feedback

* Make improvements to translated copy for the blogger importer

- Remove duplicate `<strong />` components - instead use same component
- Use variables for the importerName so that copy can be reused for
other importers
- Use siteTitle for variable name for better clarity for translators

* Use correct blogger logo on the blogger importer

- Use the blogger-alt logo from the Automattic/social-logos repo
- Set the background color of the logo on the blogger importer

* Move calls to `translate` to within the `render` method

As part of a similar change to the Medium importer, it was suggested
that the calls to `translate` should be moved within the render method
so that the strings were updated on each render of the component,
rather than translated once as the object was instantiated.

https://github.com/Automattic/wp-calypso/pull/24350#discussion_r188314191

This brings the creation of the `importer_data` object within `render`.