Post Editor: correctly reset raw content after mode switch
Wait for `setState( { mode } )` to be performed -- only then the TinyMCE is guaranteed
to be switched.
Next, save the content to post edits and reset the raw content to the value after switch
(Visual and HTML modes differ in present and missing HTML markup for the same content)
so that `initial === current`.