Refactor ellipsis-menu component style so it can be imported in WooCommerce Services
In SASS, this code:
```css
.a {
    .b & {
        color: red;
    }
}
```

Is compiled into:
```css
.b .a {
    color: red;
}
```

But when a file containing that rule is imported into a parent context, like this:
```css
.root {
    @import("a-and-b.scss");
}
```

Then the result is:
```css
.b .a .root {
    color: red;
}
```

Which is normally not what's expected. Components in `WooCommerce Services` are styled in that way, so putting an `&` in any part other than the start of a rule breaks our styles.

This PR fixes that problem. The changes should make no difference in Calypso, but by avoiding the `&` at the end of the rule it will work in `Woocommerce Services`.