Fix warning that required site.slug prop wasn't found
It could be that we no longer require this prop, but the previous
version of this code attached the `site` and `siteTitle` properties to
the active import objects. The refactor only included the `siteTitle`
and the `SiteImporterPane` component was creating a warning that a
required prop wasn't defined.

I've adapted the code to filter the active imports for the site to
include the original `map` call that appends these properties to a copy of
the object.