Make `ImporterSiteImporter` key stable to prevent remounting
The key to the `ImporterSiteImporter` component was set to the import ID
that was first set to a local value and then the ID received from the
server. This change in key was causing the `SiteImporterInputPane` to
remount and as a consequence, never sees its state change.

This fix sets the key to the importer type combined with the index of
active importers of that type.