Post Editor: improve how post edits are merged after publish date reset
This PR iterates on #25422 by improving the way how local date edits are
removed after receiving updated post from server.

Many post attributes already have nontrivial logic to compare incoming post
update with local edits and decide if the edits are applied or not. For most
attributes it's a simple `===` comparison, but for many it's not.

for `POST_SAVE_SUCCESS` action. This PR moves it to the reducer for `POSTS_RECEIVE`
action, where all the other comparisons are done.

There should be no change in functionality.