Post Editor: Remove usages of lib/user
The only usage of `lib/user` is in the "Press This" code that figures out the primary site
slug from the User object. This can be replaced safely by the `getCurrentUser` Redux selector.

We don't need to check for `null` user and wait until it's populated, like we do for sites.
The User object is always populated when logged in, as the `window.AppBoot` function will wait
until it's fetched from the network and delays the whole app boot process.