Adds Notice to warn when no more photos to show in External Media List
Previously when you reached the end of the available photos from the external media library nothing happened. Users had no feedback they’d reached the end of the list. Adds a Notice component which clearly states that there are no more photos to display.

Also upgrades the notice that appears when you’ve reached the max results upper limit of 1000.