Reader Share: clean up props of SiteSelector component
The `SiteSelector` component doesn't need the `siteBasePath` prop, because handling the
site selection is handled completely by the custom handler whose `handledByHost` return
value is always `true`. The code path that uses `siteBasePath` is never used.

Properties `showAddNewSite` and `indicator` don't need to be specified as their default
values are what we want.