Adds ability to show/hide icon for Videos
Previously icon could not be disabled. In certain situations you might need to do this. For example, Google Photos automatically embeds a video icon on to the thumbnails of all videos coming through from it’s API. As a result we need to disable our icon so as not to double up.

See
https://github.com/Automattic/wp-calypso/pull/25776#issuecomment-402199408

Originally part of PR above but moved on advice from @ehg
https://github.com/Automattic/wp-calypso/pull/25776#pullrequestreview-135410485