Creating a build script to start the Calypso SDK
See https://github.com/Automattic/wp-calypso/projects/72

This PR should not affect Calypso on WordPress.com This PR is one
component in a big project and thus represents an unfinished midpoint.
This PR is taking the first step towards reusing Calypso's libraries and
build systems as a base for building Gutenberg blocks. Our goal will be
to write blocks once in one place and have those blocks load natively
inside Calypso as well as within wp-admin

In this PR we're hijacking the Calypso build process to perform code
transpilation and build a bundle that we can import into WordPress via a
normal enqueue_script() call.