SDK: Add "editor notes" block
This patch adds a new Gutenberg block whose purpose is to provide
an area to write notes about a post while writing it while making
sure that those notes are never published in the open.

The motivation for adding a block here is to test building a block
from Calypso and then including it in another project. That is,
the output will be used in an external PHP plugin for WordPress
which loads the block itself.