Defer mounting of children in Accordion.
The accordion was designed to create its entire render tree when
mounting, and revealing/hiding it through CSS. This is a solid
approach, but it can be improved by deferring the mounting of
its children until it's actually opened for the first time.

The post editor has a sidebar which makes use of a number of
accordions, each of which with a number of components. By making
use of the approach in this PR, the amount of work in initializing
the post editor can be drastically reduced, by culling a large
portion of the render tree. The work of mounting each accordion's
children can then be performed later, if needed, when it's
actually opened, and when the main thread is likely under much
less strain.

Initial opening of the post editor takes extra time, due to other
factors, but subsequent reopens take only 38% of the total time
they used to in my machine, after this PR is applied
(303ms vs 788ms, average over 5 measurements with perfmon).