Defer the initial request on QuerySiteStats.
Requests are performed asynchronously, but have a non-trivial
setup time. Deferring the initial request allows us to schedule
that setup time for later.

In pages with a large number of QuerySiteStats, like the Stats page,
this allows the page to render earlier, improving responsiveness.

I measured a 200ms decrease (out of 750ms) in the mouseup event
handler duration when navigating to the Stats page. Total loading
time as measured by perfmon doesn't seem to be affected.