Post Editor: when saving post, overwrite the local edits with post version from server (#26247)
* Post Editor State: improve mergePostEdits to merge multiple objects

`mergePostEdits` now takes a variable number of arguments and merges them
all into one object. Also supports the "save marker" concept: string markers
in the edits log that are skipped when merging.

* Post Editor State: add new function appendToPostEditsLog

The new function appends new edits to an existing log, merging with the last one
if it's not a save marker.

* Add editorSave action

It will be used by the `saveEdited` action to insert a save marker to edit log
when save is initiated.

* Make getPostEdits process array from state and return merged edits

* Update the state.posts.edits reducer to work with edit logs and save markers

Updates the `edits` reducer to work with the new shape of post edits: edits log
instead of just one edits object.

Also adds support for adding a save marker on `EDITOR_SAVE` and removing it on
`POST_SAVE_SUCCESS`.

* Make the saveEdited action insert a save marker and remove it on success

* Post Editor: add tests for the saveEdited flow

Test the `saveEdited` flow with the real reducer and mocked network requests. Verifies
that the post edits log gets properly updated when server returns a transformed content.

* Update state/posts unit tests, mostly to reflect the changes of state.posts.edits shape