Add post editor image selector
This will act as a reusable component to allow for multiple images to be selected as featured images or for various post meta options using the media modal for image selection and uploads.

The existing editor-featured-image does this well for a single image, but does not allow multiple images to be selected and also handles data persistence directly within the component.

This new component will be in many ways similar to the featured image, but will allow multiple images and also handle data changes within a prop to allow for easy extensibility.