Domains: Don't show the last availability status for wordpress.com subdomain searches
If you search for a domain in NUX that is already taken and then you search for wordpress.com subdomain you'll get an error that the wordpress.com subdomain is already taken. This should never happen - if a wordpress.com subdomain is already taken we show a randomized variation that is available.