SDK: Explore creating preset groups for Gutenberg extensions
We already have a couple of use-cases for packaging groups of
blocks and extensions into one bundle: our internal P2 blogs
and the Jetpack extensions.

In this PR I'm collecting the P2 blocks into a new "preset"
directory with the goal of being a kind of meta-package that
imports the blocks as wanted and spits out a workable bundle.