Media Modal: Add an option to disable the gallery tools for multiple image selection (#26014)
* Add an option to disable the gallery tools for multiple image selection

The existing post editor media modal triggers a gallery view after selecting multiple images that allows for options such as tiling and linking.  Some post types (e.g., products) need to select images returned that aren't specific to the WordPress gallery.

This option will bypass the view for gallery settings and return an array of images when the action button is clicked after selecting multiple images.

* Add galleryViewEnabled prop to media modal to bypass gallery view

By default this value is true to prevent breaking all existing editor modals.  Passing in a value of false will show the 'Insert' button instead of 'Continue' that lead to the gallery settings view.