BillingReceipt: Add min-height for contenteditable (#26309)
In some browsers (eg: Chrome), `contenteditable` divs have a minimum height
automatically, but in others (eg: Firefox), that is not the case.

Normally it's possible for a user to edit the "billing details" section
of their receipts for accounting purposes. This is done via a
`contenteditable` div. The div has a hover state which shows a border
around the element.

When the receipt billing details section is empty, however, the div has
a height of 0 and therefore it has no hover state and cannot be clicked.

This change adds a `min-height` property to the element to make it
always visible.

Fixes #26114