Typography: Remove a hack for IE7 from long long ago
This ends up interfering with components that set a font size in REMs, like the Google Material components. Removing it has no impact on our work, but this removes a stumbling block to any future integrations of external components.
The interference manifests as tiny fonts, since the REM is now relative to the 62% set on html