Updates approach to retrieving folder media items
Previously a callback was passed down into MediaListData component in order to retrieve the media items that were of the “folder” type. However this approach caused a lot of misdirection and conceptual overhead. Refactored to have the MediaListData component parse the folder items out itself and pass these down to it’s child components as an additional prop alongside the original props.

See original code review comment for more info
https://github.com/Automattic/wp-calypso/pull/25763#discussion_r204868498