Migrate 'third-party' data layer handlers (refer and directly)
This PR does a few related migrations:

The `refer` data layer handler is migrated to `dispatchRequestEx`. It records a referral
to `refer.wordpress.com` when signup URL has `aff` and `cid` parameters.

The `refer` and `directly` handlers are both migrated to `registerHandlers`, so that they are moved
from `build` chunk to the chunks that use them: Signup and Inline Help.