Importers: Correct the importer type shown on the Author Mapping Pane (#26294)
* Correct the importer type shown on the Author Mapping Pane

The importer type was hardcoded as WordPress. This change passes the
current importer name down as a prop so that it can be used in the
message displayed in the map authors pane.

* Make prop name sourceType and make it optional

Addressing feedback by making other components use `sourceType` as the
prop that contains the importer name. It also makes the prop optional.