Update babel config for tests
Configures Babel for tests to transpile with the Node target. This should make
the transpilation a bit faster (fewer transforms to apply) and also help greatly
with debugging tests: the sources in debugger are the original ones, not the
transpiled ones. Async functions, for example, tend to be transpiled into code
that's hard to read for humans.

Also moves the `add-module-exports` transform to the `env.test` section.