Guided Tours: remove memoization of the hasJustSeenTour selector (#26467)
Removes a warning that says:

> Do not pass complex objects as arguments for a memoized selector

The warning is issued because an object `{ tour, _timestamp }` is passed
as selector arg.

The simplest solution is to just remove the memoization. The return value is
boolean, doesn't need to maintain identity. The selector is called only rarely,
when a tour is requested using a `?tour=` URL query param. And it's quite likely
that the memoization overhead is larger than the time avoided by doing a repeated
linear search in a small array (tour history).