Guided Tours: fix the `tourBranching` function to be able to descend render props
This patch fixes a regression in the `tourBranching` function that was introduced in #22862.
That PR converted `Step` children to a functional render prop. That led to `tourBranching`
being unable to descend into the children when examining the tour tree.

This gets fixed by `tourBranching` checking if `children` is a function, calling it when it is,
and continuing to traverse the tree using the returned children.

Additional changes:
- rename the `config-parsing.js` file to `tour-branching.js` --- more appropriate name.
- simplify the algorithm how tour branches are collected and converted into objects.
- add a unit test for the `tourBranching` function!