Guided Tours: improve API of the <Step wait=... /> function
In #22041, a new prop called `wait` was introduced to the `Step` component. It can perform
some operation (like opening a UI dialog or sidebar) that needs to be waited for before the
step can be displayed.

This patch improves its API. Right now, it takes parameters `props` and `context`, which are
arbitrary bags of arbitrary properties that expose internals of the Guided Tours system.

What we really need currently is to pass the Redux store. That's why `wait` only gets this
parameter and nothing else.

In future, we might remove even the `reduxStore` param and use the Redux Bridge instead.
Migrating Guided Tours to the new React Context API might make it difficult to make `store`
part of the tour context.