Imports: Rename tracks events for clarity
As discussed in p8F0Ux-1bF-p2 some of the tracks events recorded on the
import section could be ambiguous. This change attempts to remove that
ambiguity.

It can be tested by running an import and inspecting filtering the image
requests by `t.gif`. The associated event names and proporties can be
seen with these requests.