Use memoization to avoid re-renders in ReaderCombinedCardComponent.
Considerable time is being spent re-rendering instances of this
component both when loading a page that uses them, as well as
when navigating away from one.

This PR uses memoization in order to ensure that the existing
post keys are kept, rather than re-generated every time. This
means that we avoid unnecessary re-renders for this component and
everything below it in the render tree.

This PR, when tested in a new account, improves responsiveness
when navigating from "followed sites" to "search" by around 300ms
(before: 857, after: 538) on my machine.