Guided Tours: return only visible elements when looking for a target element
Many Guided Tours features search for a DOM element and do something with it:
- align the tour bubble to a particular DOM element
- continue or quit the tour when a DOM element is clicked
- scroll to an element to make sure it's visible
- continue or quit the tour when a target element is not found

The elements are targeted with a CSS selector that can select more than one element.
For example, in Post Editor, the "Visual Toolbar" and "HTML Toolbar" are different
components with both have an "Add" button we want to target.

This patch verifies that a selected element is visible (by checking its bounding client rect)
and returns it only if it's visible. That allows selecting the right target when the first
element is invisible and the second is.