SDK: Explore importing Calypso UI into a Gutenberg extension
In this patch I'm doing two things at once to reach the goal of using
Calypso UI components: pulling in the UI code and pulling in the styles.

As you may note I'm bringing in the entirety of the Calypso styles since
the framework styles are monolithic. This has the immediate drawback
that the stylesheet is over a megabyte in size, but the immediate
advantage that we can import Calypso UI and not have to write it all
ourselves.

If this turns out to be a good path then we will obviously have to
optimize the CSS, which is already in the plans. As the SDK and Calypso
framework improves we should see that bundled CSS drop down to very
small sizes. Currently the framework isn't ready to split apart the CSS
though I'm afraid.