Update: Stats list - refactor tests to use Jest
See #26555 (this is a preparatory commit to that change)

In order to update the way we reference flag SVGs I was facing adding
some hacks to the existing tests here on account of using `chai` instead
of `jest`. After thinking about it I decided that since we want to
eliminate `chai` anyway I may as well update the tests to use `jest`
instead and that is what I have done.

The major change is that instead of hard-coding the wanted country flag
we're only expecting it to become a string. We will test the mapping
from county code to flag SVG URL in #26555 in the new function that
performs that mapping.