Imports: Rename tracks events for clarity (#26496)
* Imports: Rename tracks events for clarity

As discussed in p8F0Ux-1bF-p2 some of the tracks events recorded on the
import section could be ambiguous. This change attempts to remove that
ambiguity.

It can be tested by running an import and inspecting filtering the image
requests by `t.gif`. The associated event names and proporties can be
seen with these requests.

* Fix `startImport` shadowing linter error

The symbol `startImport` refers to an action but is also passed in as a
prop. This caused an error with the 'no shadowing' linter rule, which
this change fixes by renaming the prop function.