SDK: Shuffle around configuration in SDK runner
When building with the SDK we have a generic runner - `sdk-cli.js`
and target-specific options for that runner - `gutenberg.js`. We
invoke the runner script while specifying the target which it calls.

Some of our configuration and runtime code though has been split
across these two places and some of the configuration could maybe
be moved upwards into the main Calypso config.

In this patch I'm shuffling around the configuration and execution
code to try and separate those layers more clearly. This work serves
to prepare us as well for integrating other projects, such as the
notifications panel, into the Calpyso repository.