Add transient image source to state for image preloader
Storing the transient image in the state allows us to pass it as a placeholder value to the image preloader and avoid the flash of images when transitioning from the blob to uploaded image.

Transient sources need to be stored in the state since we don't have access to the transient properties from the itemIds props.