Checkout: Add credit card component to replace credit card list items (#24485)
The <CreditCard> component can either render a stored card (if passed a card prop), or wrap its children (such as a new card form). Thus, it serves to replace the 'section' method of the <CreditCardSelector> component in my-sites/checkout. <StoredCard> (as used in the billing history view in me/purchases), is moved into this component directory as a child, and can continue to be imported directly into the billing history view. Documented CreditCard component and provided examples.