Imports: Add the UI for the Squarespace importer (#26609)
* Imports: Add the UI for the Squarespace importer

Although the Squarespace export is in WXR format so can be imported
using the WordPress importer, this sets up the UI so the import can be
identified as from Squarespace, and sets the stage for us to do further
preprocessing on the file.

* Rebase after importer header refactor

The `importer-header` component had a much needed refactor. This meant
that the Squarespace logo needed to be added to the options for importer
icons.