Jetpack Connect: Add additional tracking properties to the store landing page
This change allows us to include some CTA-related query args in the URL
of the store landing page and properly send them with the tracking
event that fires when the page is loaded. This will allow us to better
source conversions and better understand the origin of customers from
around jetpack.com.