Header Button: update CSS selectors to be more specific
A `HeaderButton` element has CSS class `button header-button` and the CSS selectors
`.button` and `.header-button` are currently equally specific. The one that's defined
later in the CSS stylesheet wins. That causes problems when component stylesheets get
reordered (e.g., during conversion of the build pipeline to Webpack).

This patch increases the specificity of the `.header-button` selectors and makes the
whole thing less fragile.