SDK: Add new `packages` top-level directory
As we think about pulling into this repository the sourcecode for a few
of our non-Calypso projects we need to establish a "home" for them.
In this PR we're creating a new top-level directory called `packages`
whose purpose is to hold these packages in their own namespaced
subdirectories.

Eventually we'll probably need to add special configuration options to
apply to this directory but for now it's just an empty directory with a
README.