State: Use more general-purpose wp.data plugins.
This creates a `custom-store-plugin` which is more general-purpose than
the previous `calypso-registry-plugin`. It allows any pre-existing redux
store to be used for a given wp.data store instead of wp.data creating
one itself. This allows for two things:

1. Stores can use enhancers and middleware, which is not possible in the
default wp.data configuration.
2. wp.data stores can share redux stores.

Note: This also creates an `internals-plugin`, which is necessary to
avoid the reimplementing of internals within the
`calypso-registry-plugin`. The `use` function of the `internals-plugin`
will be submitted as a PR to core Gutenberg as well so perhaps it will
not be necessary for long.