Open international labels in "production" and all other non-development environments
The feature flag for International Shipping Labels was already open in `development` and `staging`. This PR opens it in all the other environments.

This is a safe change because, if the user doesn't have a version of WooCommerce Services recent enough to support International Labels, the feature will keep being effectively disabled for them, without any error or breakage.