Media: Back Button on single media i
Currently the back button only takes the user back to the media gallery.
This PR fixes it so that the back button take the use to the previous screen.
In this case it fixes the case when the link from the activity log took the user to the meida library instead of the expected activity log.