Framework: Remove dependency on `notifications-panel` (#27081)
In #26760 we started directly importing the notifications panel code
that was merged into the Calypso repository in #26757 but I forgot to
remove the `notifications-panel` dependency from `package.json`

In this patch I'm doing just that. It should produce no visual or
functional changes to the build as right now that dependency should be
an unused dependency.