fix(state) `ui.route.query.current` has an invalid default value.
The initial state value for `ui.route.query.current` is `false`, which
turns to be invalid.

Indeed, the `MasterbarLoggedOut` component expects the current query
to be an object (`currentQuery: PropTypes.object`), and `false` is not
an object. It raised a warning, as revealed in #26958.

This patch fixes the default value of `query` to be an empty object
instead of `false`. Also, the default value with the selector
`getCurrentQueryArguments` is updated to, also, an empty object.

Tests have been updated accordingly.

Fix #26958.