Themes: back button should use previousPath if it exists
Currently the page for a single Theme will always link back to the Theme Showcase.
It is possible to navigate to a single theme page from somewhere besides the showcase.
For example, Activity Log entries link to theme pages, and in those cases, we want to
link back to the activity log.