Guided Tour Step: check for onTargetDisappear on current step on every DOM mutation
This patch fixes a crash where the Mutation Observer tries to call `onTargetDisappear`
on a `Step` that doesn't have the prop. One of previous steps had it, that's why the
Mutation Observer was initialized, but a following step doesn't need to.

The observer is removed only at the end of the tour, after all steps are executed.