Use ReactMarkdown to render content.
Previously, we were getting an HTML string from the server and setting
it blindly on the client.  With this commit, we switch to using
ReactMarkdown to render that content into a tree of React components.

This paves the way toward allowing more sophisticated documentation
content, including the use of actual components for demos.

Currently, this kills the support for highlight().  That will be
re-introduced before merge.