Theme Component: Remove overlay div (#27169)
The idea of this PR is to use HTML (well, JSX) elemens that allow us to drop the `href="javascript:;` fallback, and retain a11y functionality as required by #22861. To that end, this PR removes a `<div />` on-hover overlay from the `Theme` component, and moves its props to the remaining `<a />`.

This removes some confusion from the code (the overlay having its own `<a />` in addition to the screenshot's), and can also be seen as an effort to make it more semantic, by viewing the overlay simply as an effect rather than a navigational element. Navigation is now consolidated into the screenshot's `<a />`.