Replace multiline embeds correctly on paste
If the embed HTML that was pasted in to the editor had a blank line in
it, as is the case for some MailChimp embed codes, than the embed was
parsed in to multiple elements, and when the replacement looked at the
text content of each element in the editor DOM, the markup wouldn't be
found.

This change holds a reference to the document elements that have been
pasted in the editor and temporarily wraps a div around them so they can
be replaced easily.

The replacement is then done by finding the wrapper element, removing it
and inserting the calculated shortcode in its place.