feat(server) Start an HTTPS server (with self-signed certificate).
This patch introduces a `protocol` entry in the configuration
file. The default value is `http`, which ensures the same behaviour
than usual. Changing it to `https` and an HTTPS server will start
instead of an HTTP one. A self-signed certificate is located in
`test/server/`.

This feature is useful for Web Payment (`PaymentRequest` API, Apple
Pay, Google Pay, Samsung Pay etc.) where some API are only available
when the documents are served through HTTPS.

It has to be understood that this feature is for development only, and
probably restricted to _local_ development only. It is not aimed to be
used in production at all.