Extract function to validate reducer persistence
To simplify the function that creates a combined reducer, extract the code that validates
reducer persistence to a separate `setupReducerPersistence` function.

Also, divide `combineReducers` into two functions. In future, when dynamically adding reducers,
we will be using both of them: the one that sets up the persistence of the passed reducers first,
and the one that doesn't and assumes that the passed reducers are "final".