State: Sort store code and imports in `state/index.js`
In preparation for forthcoming work on our global state system this
patch sorts the reducers, middleware, and store enhancers in our main
state module.

Although most middleware and enhancers are using the `require()` syntax
this patch draws a new grouping of imports into `reducers`, `enhancers`,
and `middlewar` to make it easier to scan and understand the imports.

The sorting of the reducer tree is mainly for consistency.

There should be no functional, visual, or other changes to the final
build as a result of these changes.